requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
boto3
pydantic
python-dotenv